+++ src/OAuth2AdapterFactory.php
@@ -12,13 +12,11 @@ namespace Zend\Expressive\Authentication\OAuth2;
 
 use League\OAuth2\Server\ResourceServer;
 use Psr\Container\ContainerInterface;
+use Psr\Http\Message\ResponseInterface;
 use Zend\Expressive\Authentication\OAuth2\Exception;
-use Zend\Expressive\Authentication\ResponsePrototypeTrait;
 
 class OAuth2AdapterFactory
 {
-    use ResponsePrototypeTrait;
-
     public function __invoke(ContainerInterface $container) : OAuth2Adapter
     {
         $resourceServer = $container->has(ResourceServer::class)
@@ -33,7 +31,7 @@ class OAuth2AdapterFactory
 
         return new OAuth2Adapter(
             $resourceServer,
-            $this->getResponsePrototype($container)
+            $container->get(ResponseInterface::class)
         );
     }
 }