+++ test/OAuth2AdapterFactoryTest.php
@@ -12,19 +12,37 @@ namespace ZendTest\Expressive\Authentication\OAuth2;
 
 use League\OAuth2\Server\ResourceServer;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
+use stdClass;
+use TypeError;
 use Zend\Expressive\Authentication\AuthenticationInterface;
 use Zend\Expressive\Authentication\OAuth2\OAuth2Adapter;
 use Zend\Expressive\Authentication\OAuth2\OAuth2AdapterFactory;
 
 class OAuth2AdapterFactoryTest extends TestCase
 {
+    /** @var ContainerInterface|ObjectProphecy */
+    private $container;
+
+    /** @var ResourceServer|ObjectProphecy */
+    private $resourceServer;
+
+    /** @var ResponseInterface|ObjectProphecy */
+    private $response;
+
+    /** @var callable */
+    private $responseFactory;
+
     public function setUp()
     {
-        $this->container      = $this->prophesize(ContainerInterface::class);
-        $this->resourceServer = $this->prophesize(ResourceServer::class);
-        $this->response       = $this->prophesize(ResponseInterface::class);
+        $this->container       = $this->prophesize(ContainerInterface::class);
+        $this->resourceServer  = $this->prophesize(ResourceServer::class);
+        $this->response        = $this->prophesize(ResponseInterface::class);
+        $this->responseFactory = function () {
+            return $this->response->reveal();
+        };
     }
 
     public function testConstructor()
@@ -42,7 +60,7 @@ class OAuth2AdapterFactoryTest extends TestCase
         $oauth2Adapter = $factory($this->container->reveal());
     }
 
-    public function testInvokeWithResourceServerEmptyResponse()
+    public function testFactoryRaisesTypeErrorForNonCallableResponseFactory()
     {
         $this->container
             ->has(ResourceServer::class)
@@ -52,17 +70,35 @@ class OAuth2AdapterFactoryTest extends TestCase
             ->willReturn($this->resourceServer->reveal());
 
         $this->container
-            ->has(ResponseInterface::class)
-            ->willReturn(false);
+            ->get(ResponseInterface::class)
+            ->willReturn(new stdClass());
 
         $factory = new OAuth2AdapterFactory();
-        $adapter = $factory($this->container->reveal());
 
-        $this->assertInstanceOf(OAuth2Adapter::class, $adapter);
-        $this->assertInstanceOf(AuthenticationInterface::class, $adapter);
+        $this->expectException(TypeError::class);
+        $adapter = $factory($this->container->reveal());
     }
 
-    public function testInvokeResourceServerAndResponse()
+    public function testFactoryRaisesTypeErrorWhenResponseServiceProvidesResponseInstance()
+    {
+        $this->container
+            ->has(ResourceServer::class)
+            ->willReturn(true);
+        $this->container
+            ->get(ResourceServer::class)
+            ->willReturn($this->resourceServer->reveal());
+
+        $this->container
+            ->get(ResponseInterface::class)
+            ->will([$this->response, 'reveal']);
+
+        $factory = new OAuth2AdapterFactory();
+
+        $this->expectException(TypeError::class);
+        $adapter = $factory($this->container->reveal());
+    }
+
+    public function testFactoryReturnsInstanceWhenAppropriateDependenciesArePresentInContainer()
     {
         $this->container
             ->has(ResourceServer::class)
@@ -76,9 +112,7 @@ class OAuth2AdapterFactoryTest extends TestCase
             ->willReturn(true);
         $this->container
             ->get(ResponseInterface::class)
-            ->willReturn(function () {
-                return $this->response->reveal();
-            });
+            ->willReturn($this->responseFactory);
 
         $factory = new OAuth2AdapterFactory();
         $adapter = $factory($this->container->reveal());