+++ test/OAuth2AdapterTest.php
@@ -22,17 +22,29 @@ use Zend\Expressive\Authentication\UserInterface;
 
 class OAuth2AdapterTest extends TestCase
 {
+    /** @var ResourceServer|ObjectProphecy */
+    private $resourceServer;
+
+    /** @var ResponseInterface|ObjectProphecy */
+    private $response;
+
+    /** @var callable */
+    private $responseFactory;
+
     public function setUp()
     {
-        $this->resourceServer = $this->prophesize(ResourceServer::class);
-        $this->response       = $this->prophesize(ResponseInterface::class);
+        $this->resourceServer  = $this->prophesize(ResourceServer::class);
+        $this->response        = $this->prophesize(ResponseInterface::class);
+        $this->responseFactory = function () {
+            return $this->response->reveal();
+        };
     }
 
     public function testConstructor()
     {
         $adapter = new OAuth2Adapter(
             $this->resourceServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
         $this->assertInstanceOf(OAuth2Adapter::class, $adapter);
         $this->assertInstanceOf(AuthenticationInterface::class, $adapter);
@@ -50,7 +62,7 @@ class OAuth2AdapterTest extends TestCase
 
         $adapter = new OAuth2Adapter(
             $this->resourceServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $this->assertNull($adapter->authenticate($request->reveal()));
@@ -67,7 +79,7 @@ class OAuth2AdapterTest extends TestCase
 
         $adapter = new OAuth2Adapter(
             $this->resourceServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $this->assertNull($adapter->authenticate($request->reveal()));
@@ -84,7 +96,7 @@ class OAuth2AdapterTest extends TestCase
 
         $adapter = new OAuth2Adapter(
             $this->resourceServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $user = $adapter->authenticate($request->reveal());
@@ -107,7 +119,7 @@ class OAuth2AdapterTest extends TestCase
 
         $adapter = new OAuth2Adapter(
             $this->resourceServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $this->assertSame(