+++ test/OAuth2MiddlewareTest.php
@@ -15,6 +15,7 @@ use League\OAuth2\Server\Exception\OAuthServerException;
 use League\OAuth2\Server\RequestTypes\AuthorizationRequest;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
+use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamInterface;
@@ -25,20 +26,41 @@ use Zend\Expressive\Authentication\OAuth2\OAuth2Middleware;
 
 class OAuth2MiddlewareTest extends TestCase
 {
+    /** @var AuthorizationRequest|ObjectProphecy */
+    private $authRequest;
+
+    /** @var AuthorizationServer|ObjectProphecy */
+    private $authServer;
+
+    /** @var RequestHandlerInterface|ObjectProphecy */
+    private $handler;
+
+    /** @var ResponseInterface|ObjectProphecy */
+    private $response;
+
+    /** @var callable */
+    private $responseFactory;
+
+    /** @var ServerRequestInterface|ObjectProphecy */
+    private $serverRequest;
+
     public function setUp()
     {
-        $this->authServer    = $this->prophesize(AuthorizationServer::class);
-        $this->response      = $this->prophesize(ResponseInterface::class);
-        $this->serverRequest = $this->prophesize(ServerRequestInterface::class);
-        $this->authRequest   = $this->prophesize(AuthorizationRequest::class);
-        $this->handler       = $this->prophesize(RequestHandlerInterface::class);
+        $this->authServer      = $this->prophesize(AuthorizationServer::class);
+        $this->response        = $this->prophesize(ResponseInterface::class);
+        $this->serverRequest   = $this->prophesize(ServerRequestInterface::class);
+        $this->authRequest     = $this->prophesize(AuthorizationRequest::class);
+        $this->handler         = $this->prophesize(RequestHandlerInterface::class);
+        $this->responseFactory = function () {
+            return $this->response->reveal();
+        };
     }
 
     public function testConstructor()
     {
         $middleware = new OAuth2Middleware(
             $this->authServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
         $this->assertInstanceOf(OAuth2Middleware::class, $middleware);
         $this->assertInstanceOf(MiddlewareInterface::class, $middleware);
@@ -69,7 +91,7 @@ class OAuth2MiddlewareTest extends TestCase
 
         $middleware = new OAuth2Middleware(
             $this->authServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
         $response = $middleware->process(
             $this->serverRequest->reveal(),
@@ -92,7 +114,7 @@ class OAuth2MiddlewareTest extends TestCase
 
         $middleware = new OAuth2Middleware(
             $this->authServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
         $response = $middleware->process(
             $this->serverRequest->reveal(),
@@ -117,7 +139,7 @@ class OAuth2MiddlewareTest extends TestCase
 
         $middleware = new OAuth2Middleware(
             $this->authServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $this->serverRequest->getMethod()->willReturn('GET');
@@ -157,7 +179,7 @@ class OAuth2MiddlewareTest extends TestCase
 
         $middleware = new OAuth2Middleware(
             $this->authServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $this->serverRequest->getMethod()->willReturn('GET');
@@ -177,7 +199,7 @@ class OAuth2MiddlewareTest extends TestCase
 
         $middleware = new OAuth2Middleware(
             $this->authServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $response = $middleware->process(
@@ -206,7 +228,7 @@ class OAuth2MiddlewareTest extends TestCase
 
         $middleware = new OAuth2Middleware(
             $this->authServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $response = $middleware->process(
@@ -247,7 +269,7 @@ class OAuth2MiddlewareTest extends TestCase
 
         $middleware = new OAuth2Middleware(
             $this->authServer->reveal(),
-            $this->response->reveal()
+            $this->responseFactory
         );
 
         $response = $middleware->process(