+++ test/Pdo/OAuth2PdoMiddlewareTest.php
@@ -19,6 +19,7 @@ use League\OAuth2\Server\Grant\PasswordGrant;
 use League\OAuth2\Server\Grant\RefreshTokenGrant;
 use PDO;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Http\Server\RequestHandlerInterface;
 use Zend\Diactoros\Response;
 use Zend\Diactoros\ServerRequest;
@@ -39,6 +40,39 @@ class OAuth2PdoMiddlewareTest extends TestCase
     const DB_DATA        = __DIR__ . '/TestAsset/test_data.sql';
     const PRIVATE_KEY    = __DIR__ .'/../TestAsset/private.key';
     const ENCRYPTION_KEY = 'T2x2+1OGrEzfS+01OUmwhOcJiGmE58UD1fllNn6CGcQ=';
+
+    /** @var AccessTokenRepository */
+    private $accessTokenRepository;
+
+    /** @var AuthCodeRepository */
+    private $authCodeRepository;
+
+    /** @var AuthServer */
+    private $authServer;
+
+    /** @var ClientRepository */
+    private $clientRepository;
+
+    /** @var RequestHandlerInterface|ObjectProphecy */
+    private $handler;
+
+    /** @var PdoService */
+    private $pdoService;
+
+    /** @var RefreshTokenRepository */
+    private $refreshTokenRepository;
+
+    /** @var Response */
+    private $response;
+
+    /** @var callable */
+    private $responseFactory;
+
+    /** @var ScopeRepository */
+    private $scopeRepository;
+
+    /** @var UserRepository */
+    private $userRepository;
 
     public static function setUpBeforeClass()
     {
@@ -88,11 +122,17 @@ class OAuth2PdoMiddlewareTest extends TestCase
         );
 
         $this->handler = $this->prophesize(RequestHandlerInterface::class);
+        $this->responseFactory = function () {
+            return $this->response;
+        };
     }
 
     public function testConstructor()
     {
-        $authMiddleware = new OAuth2Middleware($this->authServer, $this->response);
+        $authMiddleware = new OAuth2Middleware(
+            $this->authServer,
+            $this->responseFactory
+        );
         $this->assertInstanceOf(OAuth2Middleware::class, $authMiddleware);
     }
 
@@ -123,7 +163,12 @@ class OAuth2PdoMiddlewareTest extends TestCase
             $params,
             [ 'Content-Type' => 'application/x-www-form-urlencoded' ]
         );
-        $authMiddleware = new OAuth2Middleware($this->authServer, $this->response);
+
+        $authMiddleware = new OAuth2Middleware(
+            $this->authServer,
+            $this->responseFactory
+        );
+
         $response = $authMiddleware->process($request, $this->handler->reveal());
 
         $this->assertEquals(200, $response->getStatusCode());
@@ -166,7 +211,12 @@ class OAuth2PdoMiddlewareTest extends TestCase
             $params,
             [ 'Content-Type' => 'application/x-www-form-urlencoded' ]
         );
-        $authMiddleware = new OAuth2Middleware($this->authServer, $this->response);
+
+        $authMiddleware = new OAuth2Middleware(
+            $this->authServer,
+            $this->responseFactory
+        );
+
         $response = $authMiddleware->process($request, $this->handler->reveal());
 
         $this->assertEquals(200, $response->getStatusCode());
@@ -214,7 +264,11 @@ class OAuth2PdoMiddlewareTest extends TestCase
             $params
         );
 
-        $authMiddleware = new OAuth2Middleware($this->authServer, $this->response);
+        $authMiddleware = new OAuth2Middleware(
+            $this->authServer,
+            $this->responseFactory
+        );
+
         $response = $authMiddleware->process($request, $this->handler->reveal());
 
         $this->assertEquals(302, $response->getStatusCode());
@@ -265,7 +319,12 @@ class OAuth2PdoMiddlewareTest extends TestCase
             $params,
             [ 'Content-Type' => 'application/x-www-form-urlencoded' ]
         );
-        $authMiddleware = new OAuth2Middleware($this->authServer, $this->response);
+
+        $authMiddleware = new OAuth2Middleware(
+            $this->authServer,
+            $this->responseFactory
+        );
+
         $response = $authMiddleware->process($request, $this->handler->reveal());
 
         $this->assertEquals(200, $response->getStatusCode());
@@ -307,7 +366,12 @@ class OAuth2PdoMiddlewareTest extends TestCase
             [],
             $params
         );
-        $authMiddleware = new OAuth2Middleware($this->authServer, $this->response);
+
+        $authMiddleware = new OAuth2Middleware(
+            $this->authServer,
+            $this->responseFactory
+        );
+
         $response = $authMiddleware->process($request, $this->handler->reveal());
 
         $this->assertEquals(302, $response->getStatusCode());
@@ -356,7 +420,11 @@ class OAuth2PdoMiddlewareTest extends TestCase
             [ 'Content-Type' => 'application/x-www-form-urlencoded' ]
         );
 
-        $authMiddleware = new OAuth2Middleware($this->authServer, $this->response);
+        $authMiddleware = new OAuth2Middleware(
+            $this->authServer,
+            $this->responseFactory
+        );
+
         $response = $authMiddleware->process($request, $this->handler->reveal());
 
         $this->assertEquals(200, $response->getStatusCode());